+++ src/driver/compile.mli
@@ -8,4 +8,8 @@ type linked
 
 val link : compiled list -> linked list
 
+val index : Fpath.t -> Packages.set -> unit
+
+val sherlodoc : html_dir:Fpath.t -> odoc_dir:Fpath.t -> Packages.set -> unit
+
 val html_generate : Fpath.t -> linked list -> unit