+++ src/driver/odoc_driver.ml
@@ -514,6 +514,8 @@ let run libs verbose odoc_dir html_dir stats nb_workers =
       (fun () ->
         let compiled = Compile.compile odoc_dir all in
         let linked = Compile.link compiled in
+        let () = Compile.index odoc_dir all in
+        let () = Compile.sherlodoc ~html_dir ~odoc_dir all in
         let () = Compile.html_generate html_dir linked in
         let _ = Odoc.support_files html_dir in
         ())