+++ CHANGES.md
@@ -13,6 +13,7 @@
 - OCaml 5.2.0 compatibility (@Octachron, #1094, #1112)
 - New driver package (@jonludlam, #1121)
 - Fix a big gap between the preamble and the content of a page (@EmileTrotignon, #1147)
+- Add a marshalled search index consumable by sherlodoc (@EmileTrotignon, @panglesd, #1084)
 
 ### Changed